Skip sourcemap tests in browser for now
