+++ js-api-spec/compile.browser.test.ts
@@ -5,7 +5,7 @@
 import {compileString, compileStringAsync} from 'sass';
 import type {OutputStyle} from 'sass';
 
-import {expectA as expectAsync} from './utils';
+import {expectA as expectAsync, skipForImpl} from './utils';
 
 describe('compileString', () => {
   describe('success', () => {
@@ -39,35 +39,38 @@ describe('compileString', () => {
       );
     });
 
-    describe('sourceMap', () => {
-      it("doesn't include one by default", () => {
-        expect(compileString('a {b: c}')).not.toHaveProperty('sourceMap');
-      });
-
-      it('includes one if sourceMap is true', () => {
-        const result = compileString('a {b: c}', {sourceMap: true});
-        expect(result).toHaveProperty('sourceMap');
-
-        // Explicitly don't test the details of the source map, because
-        // individual implementations are allowed to generate a custom map.
-        const sourceMap = result.sourceMap!;
-        expect(typeof sourceMap.version).toBeString();
-        expect(sourceMap.sources).toBeArray();
-        expect(sourceMap.names).toBeArray();
-        expect(sourceMap.mappings).toBeString();
-      });
-
-      it('includes one with source content if sourceMapIncludeSources is true', () => {
-        const result = compileString('a {b: c}', {
-          sourceMap: true,
-          sourceMapIncludeSources: true,
+    // Skip in browser context: https://github.com/dart-lang/sdk/issues/52200
+    skipForImpl('browser', () => {
+      describe('sourceMap', () => {
+        it("doesn't include one by default", () => {
+          expect(compileString('a {b: c}')).not.toHaveProperty('sourceMap');
         });
-        expect(result).toHaveProperty('sourceMap');
 
-        const sourceMap = result.sourceMap!;
-        expect(sourceMap).toHaveProperty('sourcesContent');
-        expect(sourceMap.sourcesContent!).toBeArray();
-        expect(sourceMap.sourcesContent!.length).toBeGreaterThanOrEqual(1);
+        it('includes one if sourceMap is true', () => {
+          const result = compileString('a {b: c}', {sourceMap: true});
+          expect(result).toHaveProperty('sourceMap');
+
+          // Explicitly don't test the details of the source map, because
+          // individual implementations are allowed to generate a custom map.
+          const sourceMap = result.sourceMap!;
+          expect(typeof sourceMap.version).toBeString();
+          expect(sourceMap.sources).toBeArray();
+          expect(sourceMap.names).toBeArray();
+          expect(sourceMap.mappings).toBeString();
+        });
+
+        it('includes one with source content if sourceMapIncludeSources is true', () => {
+          const result = compileString('a {b: c}', {
+            sourceMap: true,
+            sourceMapIncludeSources: true,
+          });
+          expect(result).toHaveProperty('sourceMap');
+
+          const sourceMap = result.sourceMap!;
+          expect(sourceMap).toHaveProperty('sourcesContent');
+          expect(sourceMap.sourcesContent!).toBeArray();
+          expect(sourceMap.sourcesContent!.length).toBeGreaterThanOrEqual(1);
+        });
       });
     });
